requirements.txt
packaging==25.0
PyYAML==6.0.3
Requests==2.32.5
